wip: properly implement manifest fetching in tag lister

+++ rust/lance/src/dataset/tag.rs
@@ -3,5 +3,5 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct TagContents {
     pub version: u64,
-    pub manifest_size: u64,
+    pub manifest_size: usize,
 }
